monir: 修改删除tab提示

+++ frontend/pc/src/views/ticket/allTicket/index.vue
@@ -105,7 +105,7 @@
                 <template>
                     <p class="bk-form-title">{{ $t(`m['基本信息']`) }}</p>
                     <bk-form-item class="bk-form-item" :label="$t(`m['自定义tab名称']`)" :required="true" :property="'name'">
-                        <bk-input v-model="customTabForm.name" :placeholder="$t(`m['请输入名称']`)"></bk-input>
+                        <bk-input v-model="customTabForm.name" :maxlength="20" :show-word-limit="true" :placeholder="$t(`m['请输入名称']`)"></bk-input>
                     </bk-form-item>
                     <bk-form-item :label="$t(`m['描述信息']`)" :required="true" :property="'desc'">
                         <bk-input v-model="customTabForm.desc" :type="'textarea'" :placeholder="$t(`m['请输入描述信息']`)"></bk-input>
@@ -444,13 +444,13 @@
                             this.getProjectTabList()
                         }
                     })
+                    this.showCustomTabEdit = false
                 }).finally(e => {
                     if (this.isEditTab) {
                         this.currentTab = this.customTabForm.name
                         this.getAllTabTicketList(this.editTabId)
                         this.editTabId = ''
                     }
-                    this.showCustomTabEdit = false
                 })
             },
             clearTabError () {
@@ -528,14 +528,19 @@
             closePanel (index, panel) {
                 // 固定tab
                 if (!this.fixedTabs.includes(panel.name)) {
-                    this.$store.dispatch('project/deleteProjectTab', this.serviceList[index].id).then(res => {
-                        if (res.result) {
-                            this.serviceList.splice(index, 1)
-                            if (this.currentTab === panel.name) {
-                                this.currentTab = '请求管理'
-                                this.serviceType = 'request'
-                            }
-                            this.getProjectTabList()
+                    this.$bkInfo({
+                        title: `请确认是否删除-[${panel.name}]`,
+                        confirmFn: () => {
+                            this.$store.dispatch('project/deleteProjectTab', this.serviceList[index].id).then(res => {
+                                if (res.result) {
+                                    this.serviceList.splice(index, 1)
+                                    if (this.currentTab === panel.name) {
+                                        this.currentTab = '请求管理'
+                                        this.serviceType = 'request'
+                                    }
+                                    this.getProjectTabList()
+                                }
+                            })
                         }
                     })
                 }
@@ -900,12 +905,11 @@
         height: 50px;
         overflow-x: auto;
         overflow-y: hidden;
-        @include scroller;
+        @include scroller(#a5a5a5, 4px, 4px);
         .drag-scroll {
             display: flex;
             justify-content: space-evenly;
             .drag-list {
-                width: 130px;
                 cursor: pointer;
                 display: inline-block;
                 height: 50px;
@@ -929,10 +933,12 @@
                     }
                 }
                 .icon-itsm-icon-three-one {
+                    width: 20px;
                     display: none;
                 }
                 .icon-edit-new {
                     display: none;
+                    width: 20px;
                 }
                 .ticket-file-count {
                     font-size: 12px;
